+++ extensions/service-catalog/tests/src/test/java/io/fabric8/servicecatalog/test/AdaptTest.java
@@ -1,0 +1,53 @@
+/**
+ * Copyright (C) 2015 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.fabric8.servicecatalog.test;
+
+import io.fabric8.kubernetes.client.DefaultKubernetesClient;
+import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.servicecatalog.client.ServiceCatalogClient;
+import io.fabric8.servicecatalog.server.mock.ServiceCatalogMockServer;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import org.junit.jupiter.migrationsupport.rules.EnableRuleMigrationSupport;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+import java.io.IOException;
+
+@EnableRuleMigrationSupport
+public class AdaptTest {
+
+  private ServiceCatalogMockServer mock = new ServiceCatalogMockServer();
+
+  @BeforeEach
+  public void setUp() {
+    mock.init();
+  }
+
+  @AfterEach
+  public void tearDown() throws IOException {
+    mock.destroy();
+  }
+
+  @Test
+  public void testAdapt() {
+    ServiceCatalogClient sc = mock.createServiceCatalog();
+    KubernetesClient kc = new DefaultKubernetesClient(sc.getConfiguration());
+    assertNotNull(kc.adapt(ServiceCatalogClient.class));
+  }
+}